revert moving holdout over to shared
