+++ packages/back-end/src/models/HoldoutModel.ts
@@ -1,4 +1,7 @@
-import { HoldoutInterface, holdoutValidator } from "shared/validators";
+import {
+  HoldoutInterface,
+  holdoutValidator,
+} from "back-end/src/validators/holdout";
 import { ExperimentInterface } from "back-end/types/experiment";
 import { MakeModelClass } from "./BaseModel";
 import { getExperimentById } from "./ExperimentModel";
