+++ packages/front-end/components/Holdout/LinkedExperimentsTable.tsx
@@ -1,4 +1,4 @@
-import { HoldoutInterface } from "shared/validators";
+import { HoldoutInterface } from "back-end/src/validators/holdout";
 import { ExperimentInterfaceStringDates } from "back-end/types/experiment";
 import { Box, Text } from "@radix-ui/themes";
 import { useRouter } from "next/router";
