+++ packages/front-end/hooks/useExperiments.ts
@@ -3,7 +3,7 @@ import {
   ExperimentType,
 } from "back-end/types/experiment";
 import { useMemo } from "react";
-import { HoldoutInterface } from "shared/validators";
+import { HoldoutInterface } from "back-end/src/validators/holdout";
 import useApi from "./useApi";
 
 export function useExperiments(
